Upgrade tests to nereid 2.4.0.8 #673
Testing was revamped in nereid 2.4.0.7 to work with unittest fixtures
and the tryton test loader. This patch migrates the module to use the
new testing and also use the travis CI engine to test the module.

Makes module depend on version 2.4.0.8 of nereid which fixes an
important issue with site-maps which is needed for this module to
generate sitemaps effectively at runtime.